feat: refactor project file
