+++ backend/src/controllers/room.controller.ts
@@ -1,9 +1,11 @@
 import { Request, Response } from "express";
-import { createRoom, createRoomPlayer } from "../models/room.model";
-import { generateRoomId, log, toError } from "../lib/utils";
-import { findUserByEmail } from "../models/user.model";
+import { createRoom, createRoomPlayer } from "../repositories/room.repository";
+import { generateRoomId, makeJoinTicket } from "../lib/utils/generators";
+import { log } from "../lib/utils/logger";
+import { toError } from "../lib/utils/errors-handlers";
+import { findUserByEmail } from "../repositories/user.repository";
 import { CurrentGameState } from "../types/game";
-import { createGameState } from "../redis/game-state";
+import { getGameState, createGameState } from "../databases/redis/game-state";
 
 type CreateRoomRequest = {
   gameType: "sam" | "phom";
@@ -134,33 +136,44 @@ export const createNewRoom = async (req: Request, res: Response) => {
   }
 };
 
-import crypto from "crypto";
-import jwt from "jsonwebtoken";
-
-const JOIN_SECRET = process.env.JOIN_SECRET || "dev-join-secret";
-const makeJoinTicket = (roomId: string, userId: number) => {
-  return jwt.sign(
-    { roomId },
-    JOIN_SECRET,
-    { subject: String(userId), expiresIn: "2m", jwtid: crypto.randomUUID() }
-  );
-}
-
-// POST /rooms/:roomId/invitations
 export const createInvitations = (req: Request, res: Response) => {
   try {
     const { roomId } = req.params as { roomId: string };
-    const { inviteeIds = [], expiresInMinutes = 30 } = (req.body ?? {}) as { inviteeIds: number[]; expiresInMinutes?: number };
+    const { inviteeIds = [], expiresInMinutes = 30 } = (req.body ?? {}) as {
+      inviteeIds: number[];
+      expiresInMinutes?: number;
+    };
+
     if (!Array.isArray(inviteeIds) || inviteeIds.length === 0) {
       return res.status(400).json({ message: "inviteeIds required" });
     }
+
     // TODO: authorize inviter (req.user.id), capacity check, upsert room_players + room_invitations
+    const gameState = getGameState(roomId);
+    if (!gameState) {
+      return res.status(404).json({ message: "Room not found" });
+    }
+
     // TODO: emit socket event 'room:invited' to each online invitee
-    return res.status(201).json({
-      roomId,
-      invitations: inviteeIds.map((uid) => ({ userId: uid, inviteId: "TBD", expiresAt: new Date(Date.now() + expiresInMinutes*60*1000).toISOString() }))
+
+
+
+    const invitations = inviteeIds.map((uid) => {
+      const expiresAt = new Date(Date.now() + expiresInMinutes * 60 * 1000);
+      const joinTicket = makeJoinTicket(roomId, uid); // 🔑 generate a short-lived ticket
+
+      return {
+        userId: uid,
+        inviteId: "TBD", // TODO: replace with DB id after insert
+        expiresAt: expiresAt.toISOString(),
+        joinTicket, // include token in response
+      };
     });
-  } catch (e: any) {
-    return res.status(500).json({ message: "Failed to create invitations", error: e?.message });
+
+    return res.status(201).json({ roomId, invitations });
+  } catch (error: unknown) {
+    const err = toError(error);
+    log("Error in createInvitations controller: ", err.message || "Internal server error", "error");
+    return res.status(500).json({ message: err.message || "Failed to create invitations"});
   }
 }
