+++ backend/src/dtos/user.dto.ts
@@ -1,0 +1,6 @@
+export type CreateUserDTO = {
+  email: string;
+  fullName: string;
+  password: string;
+  profilePic?: string;
+}
