+++ backend/src/entities/user.ts
@@ -1,0 +1,10 @@
+export interface User {
+  id: number;
+  email: string;
+  fullName: string;
+  password: string;
+  profilePic?: string;
+  balance: number;
+  createdAt: string;
+  updatedAt: string;
+}
