+++ backend/src/lib/utils/errors-handlers.ts
@@ -1,0 +1,3 @@
+export const toError = (error: unknown): Error => {
+  return error instanceof Error ? error : new Error(String(error));
+};
