+++ backend/src/routes/auth.routes.ts
@@ -6,7 +6,7 @@ import {
   logout,
   checkAuth,
 } from "../controllers/auth.controller";
-import { protectRoute } from "../middleware/auth.middleware";
+import { protectRoute } from "../middlewares/auth.middleware";
 
 const router = express.Router();
