+++ backend/src/routes/room.routes.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import { protectRoute } from "../middleware/auth.middleware";
+import { protectRoute } from "../middlewares/auth.middleware";
 import {
   createNewRoom,
   createInvitations,
