+++ backend/src/routes/user.routes.ts
@@ -1,6 +1,6 @@
 import express from "express";
 
-import { protectRoute } from "../middleware/auth.middleware";
+import { protectRoute } from "../middlewares/auth.middleware";
 import {
   updateProfile,
   getUserData,
