+++ backend/src/socket/handlers/room/join-room.handler.ts
@@ -1,8 +1,9 @@
 import { Server, Socket } from "socket.io";
-import { createRoomPlayer, findRoomByRoomId, getPlayersFromRoom, updateRoom } from "../../../models/room.model";
-import { log, toError } from "../../../lib/utils";
-import { getGameState, updateGameState } from "../../../redis/game-state";
-import { getUserById } from "../../../models/user.model";
+import { createRoomPlayer, findRoomByRoomId, getPlayersFromRoom, updateRoom } from "../../../repositories/room.repository";
+import { log } from "../../../lib/utils/logger";
+import { toError } from "../../../lib/utils/errors-handlers";
+import { getGameState, updateGameState } from "../../../databases/redis/game-state";
+import { getUserById } from "../../../repositories/user.repository";
 import { CurrentGameState } from "../../../types/game";
 
 type JoinRoomRequest = {
@@ -86,7 +87,7 @@ export const handleJoinRoom = (io: Server, socket: Socket) => {
       }
 
       await updateRoom(room);
-      await createRoomPlayer({ roomId: room.id, userId });
+      // await createRoomPlayer({ roomId: room.id, userId });
 
       log("Add user", userId, "to room:", room.id, "info");
